relays-array: fix schematics error and add ground plane
The jack connectors are all (except for the buffer input) wired
backwards, with the tip going to `GND` and the sleeve on `SIGNAL`, while
it should be the other way around.

While at it, improve a bit the PCB tracks and add a ground plane. The
latter is only connected to `GND` in a single point in order to avoid
ground loops and other unpleasant issues.